+++ internal/reconcile/reconcile_test.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2021 The Custom Pod Autoscaler Authors.
+Copyright 2022 The Custom Pod Autoscaler Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	custompodautoscalercomv1 "github.com/jthomperoo/custom-pod-autoscaler-operator/api/v1"
-	k8sreconcile "github.com/jthomperoo/custom-pod-autoscaler-operator/reconcile"
+	k8sreconcile "github.com/jthomperoo/custom-pod-autoscaler-operator/internal/reconcile"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"