+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.4.0] - 2023-08-13
 ### Added
 - Ability to pause autoscaling at a specific replica count.
 ### Changed
@@ -120,7 +122,8 @@ Autoscaler.
 - Allow creation/deletion of CPA.
 
 [Unreleased]:
-https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.3.0...HEAD
+https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.4.0...HEAD
+[v1.4.0]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.3.0...v1.4.0
 [v1.3.0]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.2.1...v1.3.0
 [v1.2.1]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.2.0...v1.2.1
 [v1.2.0]: https://github.com/jthomperoo/custom-pod-autoscaler-operator/compare/v1.1.1...v1.2.0